Games/Flame Pong/pyproject.toml
[project]
name = "flame pong"
version = "0.1.0"
description = "The classic Pong game with some added aesthetics and features!"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.6",
]
